pyproject.toml
[project]
name = "portfolio-manager"
version = "0.1.0"
description = "AI Portfolio Manager — LangGraph multi-agent, RL optimisation, RAG"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "gradio==4.44.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.40",
    # Database
    "sqlalchemy>=2.0.0",
    "pydantic>=2.7.0",
    # Visualisation
    "plotly>=5.22.0",
    # Agents
    "langgraph>=0.2.0",
    # LLM / RAG
    "langchain-core>=0.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-groq>=0.1.9",
    "langchain-chroma>=0.1.2",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.0",
    "langchain-ollama>=0.1.0",
    # Vector DB + Embeddings
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    # RL
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",
    "shimmy>=1.0.0",
    # TTS
    "gTTS>=2.5.0",
    # Knowledge base
    "wikipedia>=1.4.0",
    # MCP
    "fastmcp>=2.0.0",
    # MLOps
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "huggingface-hub>=0.23.0",
    "requests>=2.32.0",
]

[project.scripts]
portfolio = "main:main"

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
